# openapi/swagger.yaml
swagger: "2.0"
info:
  title: uuStarter Backend API (development)
  description: Provides backend functionality for to the uuStarter app.
  version: 1.0.0
host: localhost:80
schemes:
  - http
securityDefinitions:
  bearer:
    type: apiKey
    name: Authorization
    in: header
    description: >-
      Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345". 
      WARNING!! I couldn't figure out how to get auth to work in intellij idea. 
      The curl command it spits out works, but the integrated ui fails to authenticate for some reason. 
      Use the official swagger ui instead. It works there
security:
  - bearer: [ ]
definitions:
  User:
    type: object
    properties:
      email:
        type: string
        format: email
        description: The email address of the user.
      passwordHash:
        type: string
        description: The hashed password of the user.
      role:
        type: string
        enum:
          - USER
          - ADMIN
        description: The role of the user in the system.
      authType:
        type: string
        enum:
          - GOOGLE
          - BASIC
        description: The authentication method used by the user.
      userId:
        type: string
        description: Unique identifier for the user.
...
  Category:
    type: object
    properties:
      _id:
        type: string
        description: Unique identifier for the category
        example: "690b6710c001cd0566cf85a8"
      name:
        type: string
        description: Name of the category
...
paths:
...
  /users/find-by-email:
    post:
      summary: Find a user by email
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
                format: email
                description: The email address of the user to find.
      responses:
        200:
          description: User found
          schema:
            $ref: "#/definitions/User"
        404:
          description: User not found
          schema:
            type: object
            properties:
              error:
                type: string
                description: Error message indicating that the user was not found.
  /users/create-basic:
    post:
      summary: Create a new user with BASIC authentication
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
                description: The name of the user.
              email:
                type: string
                format: email
                description: The email address of the user.
              password:
                type: string
                description: The password for the user account.
      responses:
        201:
          description: User created successfully
          schema:
            $ref: '#/definitions/User'
        400:
          description: Missing name/email/password
          schema:
            type: object
            properties:
              error:
                type: string
                description: Error message indicating the missing input.
        409:
          description: Email already registered
          schema:
            type: object
            properties:
              error:
                type: string
                description: Error message indicating the email conflict.
        500:
          description: Internal error
          schema:
            type: object
            properties:
              error:
                type: string
                description: Error message for internal server error.
  /users/create-google:
    post:
      summary: Create or retrieve a user with GOOGLE authentication
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
                description: The name of the user (optional).
              email:
                type: string
                format: email
                description: The email address of the user.
      responses:
        201:
          description: User created successfully
          schema:
            $ref: '#/definitions/User'
        200:
          description: User already exists
          schema:
            $ref: '#/definitions/User'
        400:
          description: Missing email
          schema:
            type: object
            properties:
              error:
                type: string
                description: Error message indicating the missing email.
        500:
          description: Internal error
          schema:
            type: object
            properties:
              error:
                type: string
                description: Error message for internal server error.
  /users/{userId}:
    get:
      summary: "Returns a specific user"
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/User"
    put:
      summary: "Updates a specific user"
      security:
        - bearer: [ ]
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          required: true
          type: string
        - name: name
          in: body
          schema:
            type: object
            properties:
              name:
                type: string
                required: true
        - name: email
          in: body
          schema:
            type: object
            properties:
              email:
                type: string
                required: true
        - name: password
          in: body
          schema:
            type: object
            properties:
              password:
                type: string
                required: true
      responses:
        200:
          description: OK
        401:
          description: Unauthorized
        400:
          description: Failed to update
  /users/update-role/{userId}:
    put:
      summary: "Updates a specific user role"
      security:
        - bearer: [ ]
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          required: true
          type: string
        - name: role
          description: Role is enum "ADMIN","USER"
          in: body
          schema:
            type: object
            properties:
              role:
                type: string
                required: true
      responses:
        200:
          description: OK
        401:
          description: Unauthorized
        400:
          description: Failed to update

  # Auth
...
  /categories:
    get:
      summary: Returns a list of all categories
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/Category"
    post:
      summary: Adds a new category
      security:
        - bearer: [ ]
      parameters:
        - name: category
          in: body
          schema:
            type: object
            properties:
              name:
                type: string
                required: true

      produces:
        - application/json
      responses:
        200:
          description: OK
        401:
          description: Unauthorized
        400:
          description: Failed to create
  /categories/{categoryId}:
    get:
      summary: "Returns a specific category"
      produces:
        - application/json
      parameters:
        - name: categoryId
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Category"
    post:
      summary: "Updates a specific category"
      security:
        - bearer: [ ]
      produces:
        - application/json
      parameters:
        - name: categoryId
          in: path
          required: true
          type: string
        - name: category
          in: body
          schema:
            type: object
            properties:
              name:
                type: string
                required: true
      responses:
        200:
          description: OK
        401:
          description: Unauthorized
        400:
          description: Failed to update
    delete:
      summary: Delete a specific category
      security:
        - bearer: [ ]
      parameters:
        - name: categoryId
          in: path
          required: true
          type: string
          description: The unique identifier of the category.
        - name: categoryId
          in: path
          required: true
          type: string
          description: The unique identifier of the category.
      responses:
        200:
          description: Category deleted successfully
        404:
          description: Category not found
          schema:
            type: object
            properties:
              message:
                type: string
        401:
          description: Permission denied
          schema:
            type: object
            properties:
              message:
                type: string
  # Donation
...